Fix 'shift count >= width of type' compiler warning in GC_SQRT_SIZE_MAX (Cherry-pick commit 0151ecd from 'release-7_6' branch.)
* malloc.c (GC_SQRT_SIZE_MAX): Change type from unsigned to size_t.

+++ malloc.c
@@ -407,7 +407,7 @@ void * malloc(size_t lb)
 # define GC_SIZE_MAX (~(size_t)0)
 #endif
 
-#define GC_SQRT_SIZE_MAX ((1U << (WORDSZ / 2)) - 1)
+# define GC_SQRT_SIZE_MAX ((((size_t)1) << (WORDSZ / 2)) - 1)
 
 void * calloc(size_t n, size_t lb)
 {
